[Models][PointCloud][Render] Bring the camera to the coordinates' width
A load no longer forces coordinates to f4, so a cloud reaching the renderer may carry any floating width. prepare_points_for_rendering now names the coordinates' own dtype when it brings the camera onto their device, so the camera is matched to the points instead of the points being narrowed to the camera, and the depth suite pins both the render of a float64 cloud and the depth detail that survives it.

+++ models/three_d/point_cloud/render/common/prepare_points_for_rendering.py
@@ -210,7 +210,7 @@ def prepare_points_for_rendering(
     points = pc.xyz
 
     camera_prepared = camera.to(
-        device=points.device, extr_convention="opencv"
+        device=points.device, dtype=points.dtype, extr_convention="opencv"
     ).scale_intrinsics(resolution=resolution)
 
     # If `num_divide` is set, derive batch size from N / 2**num_divide.
